refactor: code cleaning and refactoring
