+++ src/lib/ai/summary.ts
@@ -1,8 +1,8 @@
 import OpenAI from 'openai';
 import { capture } from '@snapshot-labs/snapshot-sentry';
-import { fetchProposal, Proposal } from '../helpers/snapshot';
-import { IStorage } from './storage/types';
-import Cache from './cache';
+import { fetchProposal, Proposal } from '../../helpers/snapshot';
+import { IStorage } from '../storage/types';
+import Cache from '../cache';
 
 const openai = new OpenAI({ apiKey: process.env.apiKey });
 
@@ -11,7 +11,7 @@ class AISummary extends Cache {
 
   constructor(id: string, storage: IStorage) {
     super(id, storage);
-    this.filename = `snapshot-votes-report-${this.id}.csv`;
+    this.filename = `snapshot-proposal-ai-summary-${this.id}.txt`;
   }
 
   async isCacheable() {
@@ -26,16 +26,15 @@ class AISummary extends Cache {
 
   getContent = async () => {
     this.isCacheable();
+
     try {
-      const proposal = await fetchProposal(this.id);
-      const body = proposal?.body;
-      const title = proposal?.title;
-      const spaceName = proposal?.space?.name;
+      const { body, title, space } = this.proposal!;
+
       const completion = await openai.chat.completions.create({
         messages: [
           {
             role: 'system',
-            content: `The following is a governance proposal of ${spaceName}. Generate me a summary in 300 characters max. Here's the title of the proposal: '${title}' . Here's the content of the proposal: '${body}'`
+            content: `The following is a governance proposal of ${space.name}. Generate me a summary in 300 characters max. Here's the title of the proposal: '${title}' . Here's the content of the proposal: '${body}'`
           }
         ],
         model: 'gpt-4-turbo-preview'
@@ -44,13 +43,12 @@ class AISummary extends Cache {
       if (completion.choices.length === 0) {
         throw new Error('No completion in response');
       }
-      const res = completion?.choices[0];
+      const content = completion.choices[0].message.content;
 
-      if (res.message.content === null) {
+      if (!content) {
         throw new Error('No content in response');
       }
 
-      const content = res.message?.content;
       return content;
     } catch (e: any) {
       capture(e);
